orientation-event/ondeviceorientationabsolute-manual.https.html

<!DOCTYPE html>
<html>
  <head>
    <title>User agents must also provide an event handler IDL attribute named ondeviceorientationabsolute on the window object</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <p>Rotate the device to run the tests.</p>
    <button onclick="prepareAndRun(this)">Start tests</button>
    <div id="log"></div>
    <script>
      function prepareAndRun(button) {
        button.disabled = true;

        async_test(async t => {
          const permission = await DeviceOrientationEvent.requestPermission();
          t.step(() => {
            assert_equals(permission, 'granted', 'You must grant permission to run this test.');
          });

          window.ondeviceorientationabsolute = t.step_func_done((e) => {
            assert_equals(e.type, "deviceorientationabsolute",
              "Provide an event handler IDL attribute named ondeviceorientationabsolute");
            assert_true(e instanceof DeviceOrientationEvent,
              "The type of this event handler must be 'DeviceOrientationEvent");
            assert_true(e.absolute, "The absolute property must be set to true.");
          });
        }, 'ondeviceorientationabsolute event handler works');
      }
    </script>
  </body>
</html>
